Match Udacity 3.13
Move share intent to DetailActivity and match the udacity lesson 3.13
