+++ app/src/main/java/com/example/android/amountz/sunshine/app/DetailActivity.java
@@ -19,15 +19,20 @@ package com.example.android.amountz.sunshine.app;
         import android.content.Intent;
         import android.os.Bundle;
         import android.support.v4.app.Fragment;
+        import android.support.v4.view.MenuItemCompat;
         import android.support.v7.app.ActionBarActivity;
+        import android.support.v7.widget.ShareActionProvider;
+        import android.util.Log;
         import android.view.LayoutInflater;
         import android.view.Menu;
+        import android.view.MenuInflater;
         import android.view.MenuItem;
         import android.view.View;
         import android.view.ViewGroup;
         import android.widget.TextView;
 
 public class DetailActivity extends ActionBarActivity {
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -45,8 +50,11 @@ public class DetailActivity extends ActionBarActivity {
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
         getMenuInflater().inflate(R.menu.detail, menu);
+
         return true;
     }
+
+
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
@@ -68,8 +76,13 @@ public class DetailActivity extends ActionBarActivity {
      * A placeholder fragment containing a simple view.
      */
     public static class DetailFragment extends Fragment {
+        private final String LOG_TAG = DetailFragment.class.getSimpleName();
+
+        private final String FORECAST_SHARE_HASHTAG = " #SunshineApp";
+        private String mForecastStr;
 
         public DetailFragment() {
+            setHasOptionsMenu(true);
         }
 
         @Override
@@ -80,11 +93,39 @@ public class DetailActivity extends ActionBarActivity {
 
             Intent intent = getActivity().getIntent();
             if (intent != null && intent.hasExtra(Intent.EXTRA_TEXT)) {
-                String forecastText = intent.getStringExtra(Intent.EXTRA_TEXT);
-                ((TextView) rootView.findViewById(R.id.detail_text)).setText(forecastText);
+                mForecastStr = intent.getStringExtra(Intent.EXTRA_TEXT);
+                ((TextView) rootView.findViewById(R.id.detail_text)).setText(mForecastStr);
             }
-            
+
             return rootView;
+        }
+
+        @Override
+        public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+            inflater.inflate(R.menu.detailfragment, menu);
+
+            MenuItem item = menu.findItem(R.id.action_share);
+
+            ShareActionProvider mShareActionProvider =
+                    (ShareActionProvider) MenuItemCompat.getActionProvider(item);
+
+            if (mShareActionProvider != null) {
+                Log.v(LOG_TAG, "ShareActionProvider");
+                mShareActionProvider.setShareIntent(createShareForecastIntent());
+            }
+            else Log.d(LOG_TAG, "ShareActionProvider is null!!");
+
+        }
+
+        private Intent createShareForecastIntent() {
+            String shareStr = mForecastStr + FORECAST_SHARE_HASHTAG;
+            Log.v(LOG_TAG, shareStr);
+
+            Intent shareIntent = new Intent(Intent.ACTION_SEND);
+            shareIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_WHEN_TASK_RESET);
+            shareIntent.setType("text/plain");
+            shareIntent.putExtra(Intent.EXTRA_TEXT, shareStr);
+            return shareIntent;
         }
     }
 }
