+++ app/src/main/res/menu/main.xml
@@ -7,9 +7,5 @@
         android:orderInCategory="100"
         android:title="@string/action_settings"
         app:showAsAction="never" />
-    <item
-        android:id="@+id/menu_item_share"
-        android:showAsAction="always"
-        android:title="@string/action_share"
-        android:actionProviderClass="android.widget.ShareActionProvider" />
+
 </menu>
